MemoryStorage to separate module `memorystorage`.
* Factored out MemoryStorage to separate NPM module `memorystorage`
* Added devDependency on new module in package.json
* Don't expose MemoryStorage as Rhaboo.MemoryStorage
* Removed tests for MemoryStorage, they are now part of new module
* Kept tests for Rhaboo.inMemory and fallback behavior
* Added folder `node_modules` to .gitignore
