+++ .gitignore
@@ -1,3 +1,4 @@
 *.swp
 *.txt
+node_modules
