+++ manual-tests/test6.html
@@ -2,12 +2,13 @@
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
 <html>
   <script src="../rhaboo.max.js"></script>
+  <script src="../node_modules/ajon/ajon.max.js"></script>
   <link rel="stylesheet" href="http://code.jquery.com/qunit/qunit-1.14.0.css">
   <body>
     <a style="font-size:xx-large;" href="index.html">Start Again</a>
     <a style="font-size:xx-large;" href="test7.html">Next Phase</a>
     <h1>Rhaboo Tests 6</h1>
-    <h3>memoryStorage OK</h3>
+    <h3>Rhaboo.inMemory OK</h3>
     <a href="test6.js">Test code</a>
    <script src="http://code.jquery.com/qunit/qunit-1.14.0.js"></script>
     <script src="test6.js"></script>
