+++ package.json
@@ -28,7 +28,8 @@
     "grunt-browserify": "^2.1.4",
     "grunt-contrib-uglify": "^0.5.1",
     "seedrandom": "^2.3.6",
-    "parunpar": "^1.0.0"
+    "parunpar": "^1.0.0",
+	"memorystorage": "^0.9.2"
   },
   "dependencies": {
     "ajon": "^2.0.0"
